Laravel | Avançando com tinker, model e seeder

+++ back-end/desenvolvimento-web-avançado-2022-PHP-Laravel-Vue.JS/code/app_super_gestao/database/migrations/2022_08_16_201949_ajuste_produtos_filiais.php
@@ -18,7 +18,7 @@ class AjusteProdutosFiliais extends Migration
         //criando a tabela filiais
         Schema::create('filiais', function(Blueprint $table) {
             $table->id();
-            $table->string('filial',30);
+            $table->string('filial', 30);
             $table->timestamps();
         });
 
@@ -26,7 +26,7 @@ class AjusteProdutosFiliais extends Migration
         Schema::create('produto_filiais', function(Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('filial_id');
-            $table->unsignedInteger(('produto_id'));
+            $table->unsignedBigInteger(('produto_id'));
 
             $table->decimal('preco_venda', 8, 2);
             $table->integer('estoque_minimo');
